requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
django>=4.2.0,<4.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
